requirement.txt
ffmpeg
openai-whisper
deep-translator
yt-dlp